Modificacion de proyecto

+++ src/StoreAplication/Domain/Domain/Response/Product/ProductResponse.cs
@@ -7,7 +7,7 @@ namespace Domain.Response.Product
         public Guid ProductId { get; set; }
         [Required] public string Name { get; set; }
         [Required] public string Description { get; set; }
-        [Required] public int Price { get; set; }
+        [Required] public double Price { get; set; }
 
         [Required] public int InventoryStock { get; set; }
         [Required] public string Category { get; set; }
